requirements.txt
ddgs>=9.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
ollama>=0.4.0
openai>=1.40.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0